Remove adapter layer, use Koog types directly in engine
- LlmProvider interface now uses Koog Prompt, ToolDescriptor,
  Message.Response, and StreamFrame instead of custom types
- KoogLlmProvider simplified from 345 to 120 LOC (client setup
  + error mapping only, no type conversion)
- ChatEngine consumes StreamFrame directly, builds Prompt from
  ChatMessage history internally
- ToolExecutor accepts Message.Tool.Call instead of custom ToolCall
- Removed StreamEvent, LlmResult, ToolCall from LlmTypes.kt
- ChatMessage.toolCalls replaced with toolCallsJson for persistence
- Moved toToolDescriptor/compactSchema to ToolDescriptorMapping.kt

Issue: #68
