+++ app/src/main/kotlin/com/example/aapremote/assistant/llm/LlmProvider.kt
@@ -1,12 +1,14 @@
 package com.example.aapremote.assistant.llm
 
-import com.example.aapremote.assistant.engine.ChatMessage
-import com.example.aapremote.assistant.tools.ToolSpec
+import ai.koog.agents.core.tools.ToolDescriptor
+import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.message.Message
+import ai.koog.prompt.streaming.StreamFrame
 import kotlinx.coroutines.flow.Flow
 
 interface LlmProvider {
-    suspend fun generate(messages: List<ChatMessage>, tools: List<ToolSpec>, maxTokens: Int? = null): LlmResult
-    fun generateStream(messages: List<ChatMessage>, tools: List<ToolSpec>, maxTokens: Int? = null): Flow<StreamEvent>
+    suspend fun generate(prompt: Prompt, tools: List<ToolDescriptor>, maxTokens: Int? = null): List<Message.Response>
+    fun generateStream(prompt: Prompt, tools: List<ToolDescriptor>, maxTokens: Int? = null): Flow<StreamFrame>
     fun isAvailable(): Boolean
     fun modelInfo(): ModelInfo
 }
